implemented like a higher order component
