+++ .gitignore
@@ -1,4 +1,3 @@
-build
 node_modules
 npm-debug.log
 .idea/
