+++ src/index.jsx
@@ -1,0 +1,5 @@
+import Waypoint from './waypoint';
+export default Waypoint;
+
+import waypoint from './waypoint-hoc';
+export { waypoint };
